Fix crash bug when minimizing app (NSAttributedString problem)
NSAttributedString's method for instantiating from HTML seems to run in the background using a UIWebView. This has to run on the main thread. If it does not, it will eventually lead to a SIGTRAP; in our case, it's when you background the app. We're instead going to keep using DTCoreText.